backend/requirements.txt
Flask
flask-cors
SpeechRecognition
pydub
google-generativeai
gTTS
